Introducing ML detection in the Manchester decoding. BE CAREFUL!! In order to test, the ManchesterEnc is generating as output 1 and -1!!!!
git-svn-id: http://repository.cttc.es/svn/SDR/FlexiCom/trunk@347 f0b3029d-c460-4fc5-a45a-8aa6a25956dd

+++ src/layouts/VLC/bbManchesterDec.h
@@ -22,7 +22,7 @@ class bbManchesterDec : public gr_sync_decimator
 		static sptr Create(int,int);
 		int work(int, gr_vector_const_void_star &, gr_vector_void_star &);
 		~bbManchesterDec();
-		static void Decode(const int *, int *, int, int, int);
+		static void Decode(const float *, int *, int, int, int);
 	
 	private:
 		bbManchesterDec(int,int);
